+++ .vs/VSWorkspaceState.json
@@ -1,0 +1,8 @@
+{
+  "ExpandedNodes": [
+    "",
+    "\\Library"
+  ],
+  "SelectedNode": "\\Library\\Library.sln",
+  "PreviewInSolutionExplorer": false
+}
